new file:   public/log.png 	modified:   src/app/Modules/user/announcement/announcement.component.html 	modified:   src/app/Modules/user/announcement/announcement.route.ts 	modified:   src/app/Modules/user/custom-sidenav/custom-sidenav.component.ts 	modified:   src/app/Modules/user/user.module.ts 	modified:   src/app/login/login.component.ts 	modified:   src/index.html

+++ src/app/login/login.component.ts
@@ -31,7 +31,11 @@ export class LoginComponent {
     this.conn.login(this.loginForm.value).subscribe(
       (result: any) => {
         console.log('Login response:', result);
-  
+        if (result.reg_approval === null) {
+          console.warn('Registration not approved. Unable to login.');
+          alert('Your registration is pending approval. Please wait for confirmation.');
+          return; // Stop further execution
+        }
         if (result.token) {
           localStorage.setItem('token', result.token);
           localStorage.setItem('user', result.id);
